+++ src/main/java/com/wosarch/buysell/buysell/repositories/posgresql/views/TemplateAuctionsViewsRepository.java
@@ -2,7 +2,7 @@ package com.wosarch.buysell.buysell.repositories.posgresql.views;
 
 public interface TemplateAuctionsViewsRepository {
 
-    Integer getViews(String auctionSignature);
+    Long getViews(String auctionSignature);
 
-    Integer incrementViews(String auctionSignature, String remoteAddress);
+    Long incrementViews(String auctionSignature, String remoteAddress);
 }
